Change Bathrooms property type to double
Updated the Bathrooms property in PropertyMetadata from int to double to allow for fractional bathroom values (e.g., 1.5 bathrooms).

+++ src/RentalTurnManager.Models/PropertyModels.cs
@@ -30,7 +30,7 @@ public class PropertyMetadata
 {
     public string PropertyName { get; set; } = string.Empty;
     public int Bedrooms { get; set; }
-    public int Bathrooms { get; set; }
+    public double Bathrooms { get; set; }
     public string CleaningDuration { get; set; } = string.Empty;
     public string AccessInstructions { get; set; } = string.Empty;
     public string SpecialInstructions { get; set; } = string.Empty;
